Extract transaction storage to UpStorage - Sprinkle in some TODOs for future me, as a treat - Add regions in UpStorage.cs - Forget what my previous TODOs were about but leave them in anyway because next time I look at this I'll remember for sure
