+++ Nulah.Up.IntegrationTests/RuleSystemTests/RuleSystemFixture.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Nulah.Up.IntegrationTests.AccountTests;
 using Nulah.Up.IntegrationTests.Mocks;
+using Nulah.UpApi.Lib;
 using Nulah.UpApi.Lib.Controllers;
 
 namespace Nulah.Up.IntegrationTests.RuleSystemTests;
@@ -10,7 +11,7 @@ public class RuleSystemFixture : Fixture
 {
 	public TransactionController CreateController(TestUpBankApi testUpBankApi) => new TransactionController(
 		testUpBankApi,
-		Host.Services.GetRequiredService<IDocumentStore>(),
+		Host.Services.GetRequiredService<UpStorage>(),
 		new CategoryController(
 			testUpBankApi,
 			Host.Services.GetRequiredService<IDocumentStore>(),
