+++ UpApi/Nulah.UpApi.Domain/Interfaces/IUpStorage.cs
@@ -1,4 +1,6 @@
-﻿using Nulah.UpApi.Domain.Models;
+﻿using System.Linq.Expressions;
+using Nulah.UpApi.Domain.Models;
+using Nulah.UpApi.Domain.Models.Transactions;
 
 namespace Nulah.UpApi.Domain.Interfaces;
 
@@ -28,4 +30,30 @@ public interface IUpStorage
 	/// <param name="accountId"></param>
 	/// <returns></returns>
 	Task<UpAccount?> GetAccountFromCacheAsync(string accountId);
+
+	/// <summary>
+	/// Returns all transactions from the cache by given 
+	/// </summary>
+	/// <param name="pageSize"></param>
+	/// <param name="pageNumber">Defaults to 1. Must be greater than 0.</param>
+	/// <param name="queryExpression">
+	/// Defaults to ((UpTransaction)x => true) if null, returning all transactions unfiltered
+	/// </param>
+	/// <returns></returns>
+	Task<IEnumerable<UpTransaction>> LoadTransactionsFromCacheAsync(
+		int pageSize,
+		int pageNumber = 1,
+		Expression<Func<UpTransaction, bool>>? queryExpression = null);
+
+	/// <summary>
+	/// Returns a stat object for stored transactions
+	/// </summary>
+	/// <returns></returns>
+	Task<TransactionCacheStats> GetTransactionStats();
+
+	/// <summary>
+	/// Saves the given transactions to storage
+	/// </summary>
+	/// <param name="transactions"></param>
+	Task SaveTransactionsToCacheAsync(IEnumerable<UpTransaction> transactions);
 }
